backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.1
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
celery[redis]==5.4.0
redis==5.2.1
playwright==1.49.1
beautifulsoup4==4.12.3
lxml==5.3.0
httpx==0.28.1
python-multipart==0.0.20
aiofiles==24.1.0
Pillow==11.1.0
reportlab==4.2.5
